balanced evolution ending, added fast fitness and gene getting

+++ code/common.h
@@ -42,6 +42,6 @@ namespace evolution {
         double CreateFitnesses(const std::string &filename);
         int CreateNewPopulation();
         individ getGene(const int &i) { return population[i]; }
-        int evolve(const std::string &filename);
+        double evolve(const std::string &filename);
     };
 }
